+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next version
 
-- add ALIASES variable to Bot
+- add `ALIASES` variable to `Bot`
+- add `on_connected` function to `Client`
 
 ## 0.1.0 (2019-04-12)